requirements.txt
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
docker==6.1.3
dockerpty==0.4.1
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tabulate==0.9.0
termcolor==2.3.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
websocket-client==1.6.3
